+++ src/main/java/nc/apps/smartadder/dao/CategoryDAOImpl.java
@@ -28,14 +28,14 @@ public class CategoryDAOImpl implements CategoryDAO {
     public static final String SQL_GET_ID_BY_VALUES = "SELECT id from lab3_category_table where category_name = ? ";
 
     @Override
-    public Long save(Category category) throws DAOException {
+    public Integer save(Category category) throws DAOException {
         log.info("Add new author query:"+SQL_ADD_NEW);
         try(Connection con = dataSource.getConnection();
             PreparedStatement statement = con.prepareStatement(SQL_ADD_NEW)){
             statement.setString(1,category.getCategoryName());
             ResultSet res = statement.executeQuery();
             if (res.next()) {
-                return res.getLong("id");
+                return res.getInt("id");
             }else{
                 return null;
             }
@@ -45,14 +45,14 @@ public class CategoryDAOImpl implements CategoryDAO {
     }
 
     @Override
-    public Long getIdByValues(Category category) throws DAOException {
+    public Integer getIdByValues(Category category) throws DAOException {
         log.info("Add new author query:"+SQL_GET_ID_BY_VALUES);
         try(Connection con = dataSource.getConnection();
             PreparedStatement statement = con.prepareStatement(SQL_GET_ID_BY_VALUES)){
             statement.setString(1,category.getCategoryName());
             ResultSet res = statement.executeQuery();
             if (res.next()) {
-                return res.getLong("id");
+                return res.getInt("id");
             }else{
                 return null;
             }
