+++ src/main/java/nc/apps/smartadder/dao/interfaces/AuthorDAO.java
@@ -4,7 +4,7 @@ import nc.apps.smartadder.dao.exception.DAOException;
 import nc.apps.smartadder.domain.Author;
 
 public interface AuthorDAO {
-    Long save(Author author) throws DAOException;
-    Long getIdByValues(Author author) throws DAOException;
+    Integer save(Author author) throws DAOException;
+    Integer getIdByValues(Author author) throws DAOException;
 
 }
