+++ src/main/java/nc/apps/smartadder/dao/interfaces/PublisherDAO.java
@@ -4,6 +4,6 @@ import nc.apps.smartadder.dao.exception.DAOException;
 import nc.apps.smartadder.domain.Publisher;
 
 public interface PublisherDAO {
-    Long save(Publisher publisher) throws DAOException;
-    Long getIdByValues(Publisher publisher) throws DAOException;
+    Integer save(Publisher publisher) throws DAOException;
+    Integer getIdByValues(Publisher publisher) throws DAOException;
 }
